Switch to full Android.mk in-tree build
This new way of building LXC completely bypasses the native Automake
build process and uses standard Android.mk files.

This was needed to overcome some limitations of the previous
hook--honestly more like a hack--that tried to pass all the correct
toolchain flags to the native Automake system. Most importantly, the way
Android passes the C runtime libraries (crtbegin/end) depends on whether
the project is building a shared library or executable; this makes it
very hard to pass these libraries conditionally to Automake. On arm64,
LXC will not build at all, although it magically works with 32-bit arm.

This new system foregoes all of the previous toolchain flags and just
lets Android pass the flags as appropriate for the type of target being
built. The downside is that the target sources in the Android.mk files
need to be kept up-to-date with the corresponding Makefile.am files.

A few important notes:

config.h is copied over from config.h.in with hand-tuned #defines for
Android.

version.h is copied over from version.h.in and the current version of
LXC is filled out by hand.

The debian common config files are copied over from the corresponding
*.h.in file. Some of the Automake vars are not defined but they are for
non-critical parts--everything still works.

Only the bare essential components are included in the Android.mk
files.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>